Added ,checked ,Solved mistakes

+++ WebJysk/Controllers/CartController.cs
@@ -8,29 +8,61 @@ using Microsoft.AspNetCore.Authorization;
 public class CartController(ICartService cartService) : ControllerBase
 {
     private readonly ICartService service = cartService;
+    private string? GetUserId() => User.FindFirst("sub")?.Value;
+
+    [Authorize]
     [HttpPost]
-    public async Task<Response<string>> AddItemAsync(string userId, int productId, int quantity)
+    public async Task<Response<string>> AddItemAsync(int productId, int quantity)
     {
+        var userId = GetUserId();
+        if (string.IsNullOrWhiteSpace(userId))
+        {
+            return new Response<string>(System.Net.HttpStatusCode.Unauthorized, "Unauthorized");
+        }
         return await service.AddItemAsync(userId, productId, quantity);
     }
+    [Authorize]
     [HttpPut]
-    public async Task<Response<string>> UpdateQuantityAsync(string userId, int productId, int quantity)
+    public async Task<Response<string>> UpdateQuantityAsync(int productId, int quantity)
     {
-         return await service.UpdateQuantityAsync(userId, productId, quantity);
+        var userId = GetUserId();
+        if (string.IsNullOrWhiteSpace(userId))
+        {
+            return new Response<string>(System.Net.HttpStatusCode.Unauthorized, "Unauthorized");
+        }
+        return await service.UpdateQuantityAsync(userId, productId, quantity);
     }
+    [Authorize]
     [HttpDelete("remove")]
-    public async Task<Response<string>> RemoveItemAsync(string userId, int productId)
+    public async Task<Response<string>> RemoveItemAsync(int productId)
     {
+        var userId = GetUserId();
+        if (string.IsNullOrWhiteSpace(userId))
+        {
+            return new Response<string>(System.Net.HttpStatusCode.Unauthorized, "Unauthorized");
+        }
         return await service.RemoveItemAsync(userId, productId);
     }
+    [Authorize]
     [HttpDelete]
-    public async Task<Response<string>> DeltetCartAsync(string userId)
+    public async Task<Response<string>> DeltetCartAsync()
     {
+        var userId = GetUserId();
+        if (string.IsNullOrWhiteSpace(userId))
+        {
+            return new Response<string>(System.Net.HttpStatusCode.Unauthorized, "Unauthorized");
+        }
         return await service.DeltetCartAsync(userId);
     }
+    [Authorize]
     [HttpGet]
-    public async Task<Response<Cart>> GetUserCartAsync(string userId)
+    public async Task<Response<Cart>> GetUserCartAsync()
     {
+        var userId = GetUserId();
+        if (string.IsNullOrWhiteSpace(userId))
+        {
+            return new Response<Cart>(System.Net.HttpStatusCode.Unauthorized, "Unauthorized");
+        }
         return await service.GetUserCartAsync(userId);
     }
 }
